# lib/api-zod/src/generated/types/behaviorLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BehaviorLog {
  id: number;
  childId: number;
  childName: string;
  date: string;
  behavior: string;
  type: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}