# lib/api-zod/src/generated/types/child.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Child {
  id: number;
  name: string;
  age: number;
  schoolStartTime: string;
  schoolEndTime: string;
  goals: string;
  createdAt: string;
}