# lib/api-zod/src/generated/types/createBehaviorLogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBehaviorLogBody {
  childId: number;
  date: string;
  behavior: string;
  type: string;
  /** @nullable */
  notes?: string | null;
}