# lib/api-zod/src/generated/types/createRoutineBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoutineItem } from "./routineItem";

export interface CreateRoutineBody {
  childId: number;
  date: string;
  title: string;
  items: RoutineItem[];
}